🚚 Rename 'template_basic_python' → 'wordle_solver'
The template used as a base was called 'template_basic_python'. Changed
all instances of 'template_basic_python' to 'wordle_solver'.
