+++ README.md
@@ -1,13 +1,13 @@
-# template-basic-python
+# wordle-solver-python
 
 ![Coverage Report](assets/images/coverage.svg)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![linting: pylint](https://img.shields.io/badge/linting-pylint-yellowgreen)](https://github.com/PyCQA/pylint)
-[![Pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/purple-emily/template-basic-python/blob/main/.pre-commit-config.yaml)
-[![Semantic Versions](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--versions-e10079.svg)](https://github.com/template-basic-python/template-basic-python/releases)
+[![Pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/purple-emily/wordle-solver-python/blob/main/.pre-commit-config.yaml)
+[![Semantic Versions](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--versions-e10079.svg)](https://github.com/purple-emily/wordle-solver-python/releases)
 
 ## License
 
-[![License](https://img.shields.io/github/license/purple-emily/template-basic-python)](https://github.com/purple-emily/template-basic-python/blob/main/LICENSE)
+[![License](https://img.shields.io/github/license/purple-emily/wordle-solver-python)](https://github.com/purple-emily/wordle-solver-python/blob/main/LICENSE)
 
-This project is licensed under the terms of the `MIT` license. See [LICENSE](https://github.com/purple-emily/template-basic-python/blob/main/LICENSE) for more details.
+This project is licensed under the terms of the `MIT` license. See [LICENSE](https://github.com/purple-emily/wordle-solver-python/blob/main/LICENSE) for more details.
