+++ src/wordle_solver/config.py
